removed code duplication between CodeCheckService and CSharpDiagnosticService
