+++ src/OmniSharp.Roslyn.CSharp/Helpers/QuickFixExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using OmniSharp.Models;
