+++ src/OmniSharp.Roslyn.CSharp/Workers/Diagnostics/CSharpDiagnosticService.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
+using OmniSharp.Helpers;
 using OmniSharp.Models.Diagnostics;
 using OmniSharp.Roslyn;
 
@@ -139,50 +140,12 @@ namespace OmniSharp.Workers.Diagnostics
         private async Task<DiagnosticResult> ProcessNextItem(string filePath)
         {
             var documents = _workspace.GetDocuments(filePath);
-            var items = new List<DiagnosticLocation>();
-
-            if (documents.Any())
-            {
-                foreach (var document in documents)
-                {
-                    var semanticModel = await document.GetSemanticModelAsync();
-                    IEnumerable<Diagnostic> diagnostics = semanticModel.GetDiagnostics();
-
-                    foreach (var quickFix in diagnostics.Select(MakeQuickFix))
-                    {
-                        var existingQuickFix = items.FirstOrDefault(q => q.Equals(quickFix));
-                        if (existingQuickFix == null)
-                        {
-                            quickFix.Projects.Add(document.Project.Name);
-                            items.Add(quickFix);
-                        }
-                        else
-                        {
-                            existingQuickFix.Projects.Add(document.Project.Name);
-                        }
-                    }
-                }
-            }
+            var items = await documents.FindDiagnosticLocationsAsync();
 
             return new DiagnosticResult()
             {
                 FileName = filePath,
                 QuickFixes = items
-            };
-        }
-
-        private static DiagnosticLocation MakeQuickFix(Diagnostic diagnostic)
-        {
-            var span = diagnostic.Location.GetMappedLineSpan();
-            return new DiagnosticLocation
-            {
-                FileName = span.Path,
-                Line = span.StartLinePosition.Line,
-                Column = span.StartLinePosition.Character,
-                EndLine = span.EndLinePosition.Line,
-                EndColumn = span.EndLinePosition.Character,
-                Text = diagnostic.GetMessage(),
-                LogLevel = diagnostic.Severity.ToString()
             };
         }
     }
